Remove ChangeLog
The ChangeLog as it is today is not compatible with the merge-squash PR
strategy: it contains all the commits history instead of the squashed
PR. Decided to remove the ChangeLog right now. To generate a ChangeLog I
will need to search for another solution.

ChangeLog was generated by pipenv when creating the wheel distribution.