Add dbt (#17)
Added dbt integration to the project to enable SQL data transformations in database (ELT). See below details:

* Add dbt documentation

* Integrate dbt setup to project and example

* Add dbt target in Makefile

To check dbt config, run dbt models, tests and generate dbt docs

* Comment out pipenv check from CI

There is an issue with pipenv check command that check dependency graph
of the Python project. It resulted in failing CI job.

After a few google searches, someone opened an issue yesterday on pipenv
github. After reading pipenv documentation, it appears that pipenv check
uses an external service pyup.io. To use the service, pipenv uses an API
key. Issue seems that it the same for all pipenv users.

I think that pyup.io has cancelled this API key or throttled it as all
pipenv users use it. Not really sure if it is best practice doing that
in pipenv. Initially I thought it was doing some local computing on the
dependency graph.

Need to follow up the github issue (check directly the link in the
code). In the meanwhile I have decided to comment out this CI step to
avoid blocking everything.

* Add custom schema at database init

This custom schema is the one used by dbt when creating its models. The
issue is it does not configure the access privilege of the created
schema. Therefore, Metabase cannot see the views generated by Metabase.

To solve this issue, the schemas is initialized when the database is
boot using docker-compose and access privileges are configured.